update mpi boundary works, next: use mpi_prov

+++ inc/dg_boundary_table.h
@@ -13,7 +13,7 @@ struct table_elem{
 
 	int target_rank;
 
-	int coord;
+	double coord;
 
 	int hlevel;
 };
